microservice1/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-core==1.34.0
azure-identity==1.23.0
azure-keyvault-secrets==4.10.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
confluent-kafka==2.10.1
cryptography==45.0.4
fastapi==0.115.13
h11==0.16.0
idna==3.10
isodate==0.7.2
msal==1.32.3
msal-extensions==1.3.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
